+++ tools.h
@@ -77,7 +77,7 @@ void copyArray(int out[], const int in[], const int length);
 int initData();
 void initKeyboardData();
 void initTypingData();
-int compileTypingData(char *outfileName, char *const filenames[], const int multipliers[], const int length, const int unit, const int max);
+int compileTypingData(char *outfileName, char *const filenames[], const int multipliers[], const size_t length, const int unit, const size_t max);
 int sortTypingData(char **keys, int *values, int left, int right);
 char convertEscapeChar(char c);
